Add files via upload
The code uploaded contains a typical framework concerning the analysis of a text. 
In this example I used a dataset of essays both generated and not generated with AI in order to discriminate among them; in other words, my objective was to answer the question: is the text generated with AI?
In order to do so I had to find a way to synthtize my essays (each record of the dataset contains an essay) to obtain a dataset useful for machine learning analysis. 
My idea was the following: use just 15 words for each text (the most common words). 
After this pre-processing phase I implemented four different ML models (Logistic Regression, Random Forest, SGD Classifier, Naive-Bayes classifier). I obtained very good results of accuracy with all the models. 
The following step was to obtain a more balanced dataset: in fact, an unbalanced dataset can reduce the capability of prediction of the model for the less populated class (since the data pattern of this less probable class is "less learned"). To do so I used ADASYN, an oversampling technique. 
I re-estimated the four models, obtaining good results of accuracy.
At the end I implemented the entire stream for text analysis (i.e. removing stopwords, tokenizing, stemming and lemmatizing) in order to perform a second type of analysis, that is wordclouds and frequency of words.